Add files via upload
 "Improving Laplacian Pyramids Regression with Localization in Frequency and Time" was accepted for publication at ESANN 2022.
This work was written as part of my ongoing MSc in Data Science, under the supervision of Dr. Neta Fishelov Rabin from the Industrial Engineering department at Tel Aviv University.
The paper suggests two methods for improving the ALP regression. These extensions are suited to address different types of challenges in time-series forecasting tasks, like dealing with noise that is distributed non-uniformly or presents only in high frequencies of the data.